requirements.txt
flask
streamlit
pillow

